Built a context hook to share the product information to the item page.
Built ProductContext.js to create the context.

Wrapped ProductProvider around all Routes in App.js

Destructured setProductInfo from ProductContext in DisplayItem.js and made it so clicking on an item in the display would set the product info within the context.

Destructured the context from ProductContext in ItemPage.js and displayed the data.