Built a CartContext file to hold shopping cart data across components.
Built the reset button for the component via clicking on checkout button.

An item can be added to the cart via each item's respective item page.

Added the ability to increase the quantity of an item via the cart.

Added local storage for the shopping cart and item page, so that they won't reset on refresh or switching to a new page in the browser.

Fixed a bug where the quantity of items was iterating by two instead of one due to re-renders.